requirements.txt
click
numpy
torch
pillow
opencv-python
git+https://github.com/huggingface/diffusers
controlnet_aux
git+https://github.com/huggingface/transformers
torchvision
scikit-image
moviepy
tqdm
#pipeline_flux_controlnet_image_to_image.py
#pipeline_stable_diffusion_xl_img2img_reference.py
#stable_diffusion_controlnet_img2img.py
